+++ src/action_request.h
@@ -74,7 +74,7 @@ protected:
     std::string serviceID;
 
     /// \brief Client quirks
-    std::unique_ptr<Quirks> quirks;
+    std::shared_ptr<Quirks> quirks;
 
     /// \brief XML holding the response, we fill it in.
     ///
@@ -102,7 +102,7 @@ public:
     std::unique_ptr<pugi::xml_document> getRequest() const;
 
     /// \brief Returns the client quirks
-    const std::unique_ptr<Quirks>& getQuirks() const;
+    const std::shared_ptr<Quirks>& getQuirks() const;
 
     /// \brief Sets the response (XML created outside as the answer to the request)
     /// \param response XML holding the action response.