+++ src/config/setup/config_setup_boxlayout.cc
@@ -219,7 +219,7 @@ std::shared_ptr<ConfigOption> ConfigBoxLayoutSetup::newOption(const pugi::xml_no
     return optionValue;
 }
 
-std::string ConfigBoxLayoutSetup::getItemPath(const std::vector<std::size_t>& indexList, const std::vector<ConfigVal>& propOptions) const
+std::string ConfigBoxLayoutSetup::getItemPath(const std::vector<std::size_t>& indexList, const std::vector<ConfigVal>& propOptions, const std::string& propText) const
 {
     if (indexList.size() == 0) {
         if (propOptions.size() > 0) {